Split teamController: extract read + badge controllers
